modify kspace
