+++ src/force/include/cvff.h
@@ -1,8 +1,10 @@
 #pragma once
 #include "../../common/types.h"
-#include "force.h"
-#include "model/box.h"
 #include "../common/erf_table.h"
+#include "force.h"
+#include "kspace_calculator.h" //
+
+#include "model/box.h"
 #include "neighbor_list/include/neighbor_list/neighbor_list.h"
 #include "neighbor_list/include/neighbor_list_builder/full_neighbor_list_builder.h"
 class CVFF : public Force
@@ -16,56 +18,11 @@ public:
   void EvaluatePotentialenergy() override;
 
   void ComputeLJCutCoulForce();
-  void ComputeSpecialCoulForce();
   void ComputeLJVerlet();
   void ComputeLJRBL();
-
-  void ComputeKspaceForce();
-  void SumForces();
-
-  //
-  void ComputeQsqSum();
-  rbmd::Real ComputeRMS(rbmd::Id kmax,rbmd::Real box_length,rbmd::Real q2);
-  void SetKspacePara();
-
-  void  ComputeChargeStructureFactorEwald(
-          Box box,
-          rbmd::Id num_atoms,
-          Int3 Kmax_array,
-          rbmd::Real alpha,
-          rbmd::Real qqr2e,
-          thrust::host_vector<rbmd::Real> value_Re_array,
-          thrust::host_vector<rbmd::Real> value_Im_array);
-  void ComputeEwlad();//Ewald
-
-  void RBEInit(Box box,rbmd::Real alpha,rbmd::Id RBE_P);
-  void GetPsampleKey();
-  void ComputeChargeStructureFactorRBE(
-         Box box,
-         rbmd::Id num_atoms,
-         Int3 Kmax_array,
-         rbmd::Real alpha,
-         rbmd::Id RBE_P,
-         rbmd::Real qqr2e,
-         thrust::device_vector<rbmd::Real> rhok_real_redue,
-         thrust::device_vector<rbmd::Real> rhok_image_redue);
-  void ComputeRBE();//RBE
-
-
   void ComputeLJCoulEnergy();
 
-  void ComputeSelfEnergy(
-    rbmd::Real alpha,
-    rbmd::Real qqr2e,
-    rbmd::Real& ave_self_energy);  //self  Energy
-
-  void ComputeKspaceEnergy(
-        Box box,
-        rbmd::Id _num_atoms,
-        Int3 Kmax_array,
-        rbmd::Real alpha,
-        rbmd::Real qqr2e,
-        rbmd::Real& ave_ekspace);   //Ewald  Energy
+  void SumForces();
 
   void ComputeBondForce(); //Harmonic
   void ComputeAngleForce(); //Harmonic
@@ -107,36 +64,9 @@ private:
   rbmd::Real _corr_value_y = 0;
   rbmd::Real _corr_value_z = 0;
 
-  //EWALD
-  rbmd::Real _accuracy;
-  rbmd::Real _g_ewald;
-  rbmd::Real _alpha;
-  rbmd::Real _sum_sq_charge;
-  rbmd::Real* _h_Re_array;
-  rbmd::Real* _h_Im_array;
-
-  rbmd::Id _Kmax;
-  rbmd::Id _Kmax3D;
-  rbmd::Id kmax_x, kmax_y, kmax_z;
-  Int3 _kmax_array;
-  Real3 _unitk;
-  rbmd::Real _gsqmx;
-  rbmd::Id _kmax_x_orig, _kmax_y_orig, _kmax_z_orig;
-  rbmd::Id kcount;
-
-  //RBE
-  std::string _coulomb_type;
-  rbmd::Id _RBE_P;
-  std::string  _energy_rbe_flag = "yes"; //default
+  //kspace
   rbmd::Real _qqr2e;
-  rbmd::Id _num_k;
-  thrust::device_vector<rbmd::Real>  _P_Sample_x;
-  thrust::device_vector<rbmd::Real>  _P_Sample_y;
-  thrust::device_vector<rbmd::Real>  _P_Sample_z;
-  //
-  thrust::device_vector<rbmd::Id>  _psample_key;
-  thrust::device_vector<rbmd::Real> _rhok_real_redue;
-  thrust::device_vector<rbmd::Real> _rhok_image_redue;
-
+  rbmd::Real _alpha;
+  std::unique_ptr<KSpaceCalculator> _kspace_calculator;
 };
