+++ src/force/src/lj_cut_coul_kspace_op/lj_cut_coul_kspace_op.h
@@ -84,102 +84,6 @@ namespace op
                rbmd::Real* total_ecoul);
         };
 
-	template <typename DEVICE>
-	struct ComputeChargeStructureFactorOp
-	{
-		void operator()(
-			const rbmd::Id num_atoms,
-			const Real3 K,
-			const rbmd::Real* charge,
-			const rbmd::Real* px,
-			const rbmd::Real* py,
-			const rbmd::Real* pz,
-			rbmd::Real* density_real,
-			rbmd::Real* density_imag);
-	};
-
-	template <typename DEVICE>
-	struct ComputeEwaldForceOp
-	{
-		void operator()(
-			 Box box,
-			const rbmd::Id num_atoms,
-			const Int3  Kmax,
-			const rbmd::Real alpha,
-			const rbmd::Real qqr2e,
-			const rbmd::Real* real_array,
-			const rbmd::Real* imag_array,
-			const rbmd::Real* charge,
-			const rbmd::Real* px,
-			const rbmd::Real* py,
-			const rbmd::Real* pz,
-			rbmd::Real* fx,
-                        rbmd::Real* fy,
-                        rbmd::Real* fz,
-                        rbmd::Real* flat_virial);
-	};
-
-       template<typename DEVICE>
-       struct SqchargeOp
-        {
-         void operator()(
-           const rbmd::Id num_atoms,
-           const rbmd::Real* charge,
-           rbmd::Real* sq_charge);
-       };
-
-
-        template<typename DEVICE>
-        struct GenerateIndexArrayOp
-        {
-          void operator()(
-          const rbmd::Id  num_atoms,
-          const rbmd::Id  RBE_P,
-          rbmd::Id* psample_key);
-        };
-
-       //RBE
-       template <typename DEVICE>
-       struct ComputePnumberChargeStructureFactorOp
-       {
-         void operator()(
-             Box box,
-            const rbmd::Id num_atoms,
-            const rbmd::Id p_number,
-            const rbmd::Real* charge,
-            const rbmd::Real* p_sample_x,
-            const rbmd::Real* p_sample_y,
-            const rbmd::Real* p_sample_z,
-            const rbmd::Real* px,
-            const rbmd::Real* py,
-            const rbmd::Real* pz,
-            rbmd::Real* density_real,
-            rbmd::Real* density_imag);
-       };
-
-      template <typename DEVICE>
-      struct ComputeRBEForceOp
-      {
-        void operator()(
-            Box box,
-           const rbmd::Id num_atoms,
-           const rbmd::Id  p_number,
-           const rbmd::Real alpha,
-           const rbmd::Real qqr2e,
-           const rbmd::Real* real_array,
-           const rbmd::Real* imag_array,
-           const rbmd::Real* charge,
-           const rbmd::Real* p_sample_x,
-           const rbmd::Real* p_sample_y,
-           const rbmd::Real* p_sample_z,
-           const rbmd::Real* px,
-           const rbmd::Real* py,
-           const rbmd::Real* pz,
-           rbmd::Real* fx,
-           rbmd::Real* fy,
-           rbmd::Real* fz,
-           rbmd::Real* flat_virial);
-      };
 
       template <typename DEVICE>
       struct AddForceOp
@@ -194,50 +98,9 @@ namespace op
                 rbmd::Real* fz);
       };
 
-    template <typename DEVICE>
-    struct EikOp
-    {
-      void operator()(
-      const rbmd::Id num_atoms,const rbmd::Real gsqmx,Real3 unitk,
-      const rbmd::Id kmax,Int3 kmax_array,const rbmd::Real* px,
-      const rbmd::Real* py,const rbmd::Real* pz,const rbmd::Real* charge,
-      rbmd::Real* cs, rbmd::Real* sn,rbmd::Real* sfacrl, rbmd::Real* sfacim);
-    };
 
-    template <typename DEVICE>
-    struct EikFixOP
-    {
-      void operator()(
-      const rbmd::Id num_atoms, const rbmd::Id  total_wavevectors,
-      const rbmd::Real gsqmx,Real3 *unitk,
-      const rbmd::Id* kxvecs,const rbmd::Id *kyvecs,const rbmd::Id *kzvecs,
-      const rbmd::Id* atoms_id,const rbmd::Id* atom_id_to_idx,
-      const rbmd::Real* px,const rbmd::Real* py,const rbmd::Real* pz,
-      const rbmd::Real* charge,rbmd::Real *sfacrl, rbmd::Real *sfacim);
-    };
 
-  template <typename DEVICE>
-  struct EikFix3OP
-  {
-      void operator()(
-      const rbmd::Id num_atoms, const rbmd::Id  total_wavevectors,
-      const rbmd::Real* kxvecs,const rbmd::Real* kyvecs,const rbmd::Real* kzvecs,
-      const rbmd::Id* atoms_id,const rbmd::Id* atom_id_to_idx,
-      const rbmd::Real* px,const rbmd::Real* py,const rbmd::Real* pz,
-      const rbmd::Real* charge,rbmd::Real *sfacrl, rbmd::Real *sfacim);
-  };
 
-    template <typename DEVICE>
-    struct EwaldForceFixOp
-    {
-      void operator()(
-      const rbmd::Id num_atoms,const rbmd::Id kcount,const rbmd::Id k_index,
-      const rbmd::Real qqr2e,Int3 kmax_vec3D,const rbmd::Real* eg,
-      const rbmd::Real* cs,const rbmd::Real* sn,const rbmd::Real* charge,
-      const rbmd::Real* qfactor_real,const rbmd::Real* qfactor_image,
-      rbmd::Real* fx, rbmd::Real* fy, rbmd::Real* fz);
-
-    };
         // // // // // // // // // // // // // // // // // // //
         template <>
         struct LJCutCoulForceOp<device::DEVICE_GPU>
@@ -318,105 +181,6 @@ namespace op
        };
 
 
-
-	template <>
-	struct ComputeChargeStructureFactorOp<device::DEVICE_GPU>
-	{
-		void operator()(
-			const rbmd::Id num_atoms,
-			const Real3 K,
-			const rbmd::Real* charge,
-			const rbmd::Real* px,
-			const rbmd::Real* py,
-			const rbmd::Real* pz,
-			rbmd::Real* density_real,
-			rbmd::Real* density_imag);
-	};
-
-	template <>
-	struct ComputeEwaldForceOp<device::DEVICE_GPU>
-	{
-		void operator()(
-			 Box box,
-			const rbmd::Id num_atoms,
-			const Int3  Kmax,
-			const rbmd::Real alpha,
-			const rbmd::Real qqr2e,
-			const rbmd::Real* real_array,
-			const rbmd::Real* imag_array,
-			const rbmd::Real* charge,
-			const rbmd::Real* px,
-			const rbmd::Real* py,
-			const rbmd::Real* pz,
-			rbmd::Real* fx,
-                        rbmd::Real* fy,
-                        rbmd::Real* fz,
-                        rbmd::Real* flat_virial);
-	};
-
-        template<>
-        struct SqchargeOp<device::DEVICE_GPU>
-        {
-          void operator()(
-            const rbmd::Id num_atoms,
-            const rbmd::Real* charge,
-            rbmd::Real* sq_charge);
-        };
-
-
-        //
-       template<>
-       struct GenerateIndexArrayOp<device::DEVICE_GPU>
-       {
-         void operator()(
-         const rbmd::Id  num_atoms,
-         const rbmd::Id  RBE_P,
-         rbmd::Id* psample_key);
-       };
-
-        //RBE
-       template <>
-       struct ComputePnumberChargeStructureFactorOp<device::DEVICE_GPU>
-       {
-         void operator()(
-             Box box,
-            const rbmd::Id num_atoms,
-            const rbmd::Id p_number,
-            const rbmd::Real* charge,
-            const rbmd::Real* p_sample_x,
-            const rbmd::Real* p_sample_y,
-            const rbmd::Real* p_sample_z,
-            const rbmd::Real* px,
-            const rbmd::Real* py,
-            const rbmd::Real* pz,
-            rbmd::Real* density_real,
-            rbmd::Real* density_imag);
-       };
-
-      template <>
-      struct ComputeRBEForceOp<device::DEVICE_GPU>
-      {
-        void operator()(
-            Box box,
-           const rbmd::Id num_atoms,
-           const rbmd::Id  p_number,
-           const rbmd::Real alpha,
-           const rbmd::Real qqr2e,
-           const rbmd::Real* real_array,
-           const rbmd::Real* imag_array,
-           const rbmd::Real* charge,
-           const rbmd::Real* p_sample_x,
-           const rbmd::Real* p_sample_y,
-           const rbmd::Real* p_sample_z,
-           const rbmd::Real* px,
-           const rbmd::Real* py,
-           const rbmd::Real* pz,
-           rbmd::Real* fx,
-           rbmd::Real* fy,
-           rbmd::Real* fz,
-           rbmd::Real* flat_virial);
-      };
-
     template <>
     struct AddForceOp<device::DEVICE_GPU>
     {
@@ -429,50 +193,5 @@ namespace op
               rbmd::Real* fy,
               rbmd::Real* fz);
     };
-
-  template <>
-  struct EikOp<device::DEVICE_GPU>
-  {
-    void operator()(
-    const rbmd::Id num_atoms,const rbmd::Real gsqmx,Real3 unitk,
-    const rbmd::Id kmax,Int3 kmax_array, const rbmd::Real* px,
-    const rbmd::Real* py,const rbmd::Real* pz,const rbmd::Real* charge,
-    rbmd::Real* cs, rbmd::Real* sn,rbmd::Real* sfacrl, rbmd::Real* sfacim);
-  };
-
-  template <>
-  struct EikFixOP<device::DEVICE_GPU>{
-    void operator()(
-    const rbmd::Id num_atoms, const rbmd::Id total_wavevectors,
-    const rbmd::Real gsqmx,Real3 unitk,
-    const rbmd::Id* kxvecs,const rbmd::Id *kyvecs,const rbmd::Id *kzvecs,
-    const rbmd::Id* atoms_id,const rbmd::Id* atom_id_to_idx,
-    const rbmd::Real* px,const rbmd::Real* py,const rbmd::Real* pz,
-    const rbmd::Real* charge,rbmd::Real *sfacrl, rbmd::Real *sfacim);
-  };
-
-  template <>
-  struct EikFix3OP<device::DEVICE_GPU> {
-    void operator()(
-    const rbmd::Id num_atoms, const rbmd::Id  total_wavevectors,
-    const rbmd::Real* kxvecs,const rbmd::Real* kyvecs,const rbmd::Real* kzvecs,
-    const rbmd::Id* atoms_id,const rbmd::Id* atom_id_to_idx,
-    const rbmd::Real* px,const rbmd::Real* py,const rbmd::Real* pz,
-    const rbmd::Real* charge,rbmd::Real *sfacrl, rbmd::Real *sfacim);
-  };
-
-
-  template <>
-  struct EwaldForceFixOp<device::DEVICE_GPU>
-  {
-    void operator()(
-    const rbmd::Id num_atoms,const rbmd::Id kcount, const rbmd::Id k_index,
-    const rbmd::Real qqr2e, Int3 kmax_vec3D,const rbmd::Real* eg,
-    const rbmd::Real* cs,const rbmd::Real* sn,const rbmd::Real* charge,
-    const rbmd::Real* qfactor_real,const rbmd::Real* qfactor_image,
-    rbmd::Real* fx, rbmd::Real* fy, rbmd::Real* fz);
-
-  };
-
 
 }// namespace op
