+++ src/simulate_pipeline/src/controller/velocity_controller/default_velocity_controller.cpp
@@ -57,14 +57,4 @@ void DefaultVelocityController::Update() {
       thrust::raw_pointer_cast(_device_data->_d_vy.data()),
       thrust::raw_pointer_cast(_device_data->_d_vz.data()));
 
-  // op::UpdateVelocityOp<device::DEVICE_GPU>()(
-  //   *(_structure_info_data->_num_atoms), _dt, _fmt2v,
-  //   thrust::raw_pointer_cast(_device_data->_d_atoms_type.data()),
-  //   thrust::raw_pointer_cast(_device_data->_d_mass.data()),
-  //   thrust::raw_pointer_cast(_device_data->_d_fx.data()),
-  //   thrust::raw_pointer_cast(_device_data->_d_fy.data()),
-  //   thrust::raw_pointer_cast(_device_data->_d_fz.data()),
-  //   thrust::raw_pointer_cast(_device_data->_d_vx.data()),
-  //   thrust::raw_pointer_cast(_device_data->_d_vy.data()),
-  //   thrust::raw_pointer_cast(_device_data->_d_vz.data()));
 }
